+++ README.md
@@ -10,13 +10,13 @@
 
 `npm run start:dev`
 
-* App served @ `http://localhost:8181` with nodemon
+* App served @ `http://localhost:8080` with nodemon
 
 **Production**
 
 `npm run start`
 
-* App served @ `http://localhost:8181` without nodemon
+* App served @ `http://localhost:8080` without nodemon
 
 ---
 
@@ -24,7 +24,7 @@
 
 Command | Description
 --- | ---
-`npm run dev` | App served @ `http://localhost:8181` with nodemon
-`npm run start` | App served @ `http://localhost:8181` without nodemon
+`npm run dev` | App served @ `http://localhost:8080` with nodemon
+`npm run start` | App served @ `http://localhost:8080` without nodemon
 
 **Note**: replace `npm` with `yarn` in `package.json` if you use yarn.
