+++ src/drawing.ts
